memegen/requirements.txt
# Core dependencies
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# CLI and I/O
click>=8.1.0
pillow>=10.0.0

# Testing
pytest>=7.4.0

# Optional: sentence-transformers for better RAG
# Uncomment to enable:
# sentence-transformers>=2.2.0

# Optional: for plotting in eval
matplotlib>=3.7.0
